introduced ItemView, ItemViewHolder and ListItem
